apps/api/pyproject.toml
[project]
name = "agent"
version = "0.0.1"
description = "Agent powered by LangGraph"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "composio>=0.11.1",
    "hyperbrowser>=0.39.0",
    "composio-langchain>=0.11.1",
    "deepagents>=0.4.1",
    "fastapi>=0.115.0",
    "langchain>=1.2.10",
    "langchain-anthropic>=1.3.3",
    "langchain-community>=0.3.0",
    "langchain-modal>=0.0.1",
    "langgraph>=1.0.8",
    "langgraph-checkpoint-postgres>=3.0.4",
    "langgraph-cli[inmem]>=0.4.12",
    "langsmith>=0.1.0",
    "langmem>=0.0.30",
    "llamafirewall>=1.0.3",
    "playwright>=1.49.0",
    "psycopg[binary,pool]>=3.3.2",
    "rich>=13.7.0",
    "beautifulsoup4>=4.14.3",
    "ipython>=9.10.0",
    "uvicorn>=0.34.0",
    "matplotlib>=3.9.0",
    "networkx>=3.6.1",
    "neo4j>=6.1.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "langgraph-cli>=0.2.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
